1.小完善， 2.增加restapp.ReadJson方法，解决中文日期格式反序列化问题。不需要使用types.JSONTime类型。

+++ readexcel/read_file_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 )
 
-func testReviewFile(t *testing.T, file string, sheetName string, maxRows int) {
+func testReviewFile(t *testing.T, file string, sheetName string, maxRows int64) {
 	review, err := ReviewFile(file, sheetName, maxRows)
 	h := ""
 	for i := 0; i < len(review.Columns); i++ {
